refactor(frontend): move collections out of media
