+++ apps/frontend/src/pages/collections/index.tsx
@@ -21,7 +21,7 @@ import { useQuery } from "@tanstack/react-query";
 import Head from "next/head";
 import { useRouter } from "next/router";
 import { type ReactElement } from "react";
-import type { NextPageWithLayout } from "../../_app";
+import type { NextPageWithLayout } from "../_app";
 
 const Page: NextPageWithLayout = () => {
 	const router = useRouter();
